Backend now chooses correct database to use
Added an environment variable in backend/Dockerfile to decide whether
we're in Docker or not. (DB_BACKEND_IN_DOCKER=1)

Then, this environment variable is examined in settings.py.

If it's not set, we use the local database as normal.

If it is set, then we know we have access to Postgres, so we use that.